Move hospital members' definitions outside hospital class
